requirements.txt
flask
rembg[gpu]
pillow
onnxruntime
gunicorn